Move some stuff to the bot cog and update some dates
